rename buffer_vec.rs -> byte_buffer.rs
